requirements.txt
Flask==3.0.2
WeasyPrint==60.2
pydyf==0.8.0
gunicorn==21.2.0
requests==2.32.3
pyjwt==2.10.1
cryptography==45.0.3
python-dotenv==1.1.0